+++ custom_components/stromer/__init__.py
@@ -10,14 +10,15 @@ from homeassistant.helpers import device_registry as dr
 
 from .const import CONF_CLIENT_ID, CONF_CLIENT_SECRET, DOMAIN, LOGGER
 from .coordinator import StromerDataUpdateCoordinator
-from .stromer import Stromer, ApiError
+from .stromer import ApiError, Stromer
 
 SCAN_INTERVAL = timedelta(minutes=10)
 
 PLATFORMS: list[Platform] = [
-    Platform.SENSOR,
     Platform.BINARY_SENSOR,
     Platform.DEVICE_TRACKER,
+    Platform.SENSOR,
+    Platform.SWITCH,
 ]
 
 
@@ -45,7 +46,7 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         hass.config_entries.async_update_entry(entry, unique_id=stromer.bike_id)
 
     # Set up coordinator for fetching data
-    coordinator = StromerDataUpdateCoordinator(hass, stromer, SCAN_INTERVAL)
+    coordinator = StromerDataUpdateCoordinator(hass, stromer, SCAN_INTERVAL)  # type: ignore[arg-type]
     await coordinator.async_config_entry_first_refresh()
 
     # Store coordinator for use in platforms
@@ -72,4 +73,4 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     if unload_ok := await hass.config_entries.async_unload_platforms(entry, PLATFORMS):
         hass.data[DOMAIN].pop(entry.entry_id)
 
-    return unload_ok
+    return unload_ok  # type: ignore [no-any-return]