+++ custom_components/stromer/binary_sensor.py
@@ -7,16 +7,18 @@ from homeassistant.components.binary_sensor import (
     BinarySensorEntity,
     BinarySensorEntityDescription,
 )
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity import EntityCategory
+from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from custom_components.stromer.coordinator import StromerDataUpdateCoordinator
-
-from .const import DOMAIN
+from .const import DOMAIN, LOGGER
+from .coordinator import StromerDataUpdateCoordinator
 from .entity import StromerEntity
 
 
 @dataclass
-class StromerBinarySensorEntityDescription(BinarySensorEntityDescription):
+class StromerBinarySensorEntityDescription(BinarySensorEntityDescription):  # type: ignore[misc]
     """Describes a Stromer binary sensor entity."""
 
     icon_off: str | None = None
@@ -25,21 +27,21 @@ class StromerBinarySensorEntityDescription(BinarySensorEntityDescription):
 BINARY_SENSORS: tuple[StromerBinarySensorEntityDescription, ...] = (
     StromerBinarySensorEntityDescription(
         key="light_on",
-        name="Light on",
+        translation_key="light_on",
         icon="mdi:lightbulb",
         icon_off="mdi:lightbulb-off",
         entity_category=EntityCategory.DIAGNOSTIC,
     ),
     StromerBinarySensorEntityDescription(
         key="lock_flag",
-        name="Bike Lock",
+        translation_key="lock_flag",
         icon="mdi:lock",
         icon_off="mdi:lock-open",
         entity_category=EntityCategory.DIAGNOSTIC,
     ),
     StromerBinarySensorEntityDescription(
         key="theft_flag",
-        name="Theft flag",
+        translation_key="theft_flag",
         icon="mdi:alarm-light",
         icon_off="mdi:shield-moon",
         entity_category=EntityCategory.DIAGNOSTIC,
@@ -47,7 +49,7 @@ BINARY_SENSORS: tuple[StromerBinarySensorEntityDescription, ...] = (
 )
 
 
-async def async_setup_entry(hass, config_entry, async_add_entities):
+async def async_setup_entry(hass: HomeAssistant, config_entry: ConfigEntry, async_add_entities: AddEntitiesCallback) -> None:
     """Set up the Stromer sensors from a config entry."""
     coordinator = hass.data[DOMAIN][config_entry.entry_id]
 
@@ -58,12 +60,19 @@ async def async_setup_entry(hass, config_entry, async_add_entities):
                 entities.append(
                     StromerBinarySensor(coordinator, idx, data, description)
                 )
+                LOGGER.debug(
+                    "Add %s %s binary_sensor", data, description.translation_key
+                )
 
     async_add_entities(entities, update_before_add=False)
 
 
-class StromerBinarySensor(StromerEntity, BinarySensorEntity):
+class StromerBinarySensor(StromerEntity, BinarySensorEntity):  # type: ignore[misc]
     """Representation of a Binary Sensor."""
+
+    _attr_has_entity_name = True
+
+    entity_description = StromerBinarySensorEntityDescription
 
     def __init__(
         self,
@@ -74,18 +83,15 @@ class StromerBinarySensor(StromerEntity, BinarySensorEntity):
     ):
         """Initialize the sensor."""
         super().__init__(coordinator)
-        self._idx = idx
         self._ent = data[0]
-        self._data = data[1]
         self._coordinator = coordinator
 
         device_id = coordinator.data.bike_id
 
         self.entity_description = description
         self._attr_unique_id = f"{device_id}-{description.key}"
-        self._attr_name = (f"{coordinator.data.bike_name} {description.name}").lstrip()
 
     @property
     def is_on(self) -> bool | None:
         """Return true if the binary sensor is on."""
-        return self._coordinator.data.bikedata.get(self._ent)
+        return self._coordinator.data.bikedata.get(self._ent)  # type: ignore[no-any-return]