+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Changelog
 
+### DEC 2023 [0.3.1 (alpha/beta]
+
+- Released version 0.3
+
+### DEC 2023 [0.3.0 (alpha/beta]
+
+- Introduce Lock and Light switches
+- Fix UnitOf[X] for future HA compatibility (e.g. TEMP_CELSIUS => UnitOfTemperature.CELSIUS)
+- Improve naming through translation_keys (as also introduced way back)
+
 ### DEC 2023 [0.2.9]
 
 - Fix energy measurement to total/total_increasing